requirements.txt
# DonTouchMe 项目依赖

# 核心功能
opencv-python>=4.8.0
mss>=9.0.1
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0

# GUI
PyQt5>=5.15.0

# Windows 集成（后续阶段使用）
# pywin32>=306

# 加密（后续阶段使用）
# cryptography>=41.0.0

# 任务调度备用
schedule>=1.2.0
